[Cli] Move Provider class to abstract namespace.
