+++ src/Valkyrja/Cli/Routing/Provider/CliRouteProvider.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Valkyrja\Cli\Routing\Provider;
 
 use Override;
+use Valkyrja\Cli\Routing\Provider\Abstract\Provider;
 use Valkyrja\Cli\Server\Command\HelpCommand;
 use Valkyrja\Cli\Server\Command\ListBashCommand;
 use Valkyrja\Cli\Server\Command\ListCommand;
