+++ tests/Tests/Classes/Cli/Routing/Provider/ProviderClass.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Tests\Classes\Cli\Routing\Provider;
 
-use Valkyrja\Cli\Routing\Provider\Provider;
+use Valkyrja\Cli\Routing\Provider\Abstract\Provider;
 
 final class ProviderClass extends Provider
 {
